removed empty files
