--- changes/signals.py
@@ -1,6 +1,0 @@
-def psh_validate_implement(sender, instance, created, **kwargs):
-    """Post save hook to fire Change validation task"""
-    if created:
-        from .tasks import validate_implement
-
-        validate_implement.apply_async(kwargs={"change_id": str(instance.id)})
